move no-gateways content into new component
