+++ src/nextapp/components/gateway-get-started/index.ts
@@ -1,0 +1,1 @@
+export { default } from './gateway-get-started';
